+++ edgehog-device-runtime-containers/CHANGELOG.md
@@ -25,6 +25,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project
 adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+
+- (containers): keep track of the deployment's container id index received from Astarte
+
 ## [0.10.6] - 2026-05-13
 
 ## [0.10.5] - 2026-04-07